ContactApp/backend/requirements.txt
Django==2.2.3
djangorestframework==3.9.4
django-cors-headers==3.0.2
django-ratelimit==2.0.0